Added Keyword and KeywordLists to OData

+++ clownfish/src/main/java/io/clownfish/clownfish/odata/EntityUtil.java
@@ -586,14 +586,16 @@ public class EntityUtil {
                         }
                     };
                     if (canCreate) {
-                        ArrayList al = (ArrayList) requestEntity.getProperty("keywordset").getValue();
-                        for (Object o : al) {
-                            CfClasscontentkeyword cckw = new CfClasscontentkeyword();
-                            CfClasscontentkeywordPK cfClasscontentkeywordPK = new CfClasscontentkeywordPK();
-                            cfClasscontentkeywordPK.setClasscontentref(newclasscontent2.getId());
-                            cfClasscontentkeywordPK.setKeywordref(Long.parseLong(o.toString()));
-                            cckw.setCfClasscontentkeywordPK(cfClasscontentkeywordPK);
-                            cfclasscontentkeywordService.create(cckw);
+                        if (null != requestEntity.getProperty("keywordset")) {
+                            ArrayList al = (ArrayList) requestEntity.getProperty("keywordset").getValue();
+                            for (Object o : al) {
+                                CfClasscontentkeyword cckw = new CfClasscontentkeyword();
+                                CfClasscontentkeywordPK cfClasscontentkeywordPK = new CfClasscontentkeywordPK();
+                                cfClasscontentkeywordPK.setClasscontentref(newclasscontent2.getId());
+                                cfClasscontentkeywordPK.setKeywordref(Long.parseLong(o.toString()));
+                                cckw.setCfClasscontentkeywordPK(cfClasscontentkeywordPK);
+                                cfclasscontentkeywordService.create(cckw);
+                            }
                         }
                         hibernateUtil.insertContent(newclasscontent);
                         contentUtil.commit(newclasscontent);
@@ -903,19 +905,21 @@ public class EntityUtil {
                             }
                         }
                         
-                        // Delete corresponding keywordcontent entries
-                        List<CfClasscontentkeyword> keywordcontentdummy = cfclasscontentkeywordService.findByClassContentRef(cfclasscontent.getId());
-                        for (CfClasscontentkeyword keywordcontent : keywordcontentdummy) {
-                            cfclasscontentkeywordService.delete(keywordcontent);
-                        }
-                        ArrayList al = (ArrayList) entity.getProperty("keywordset").getValue();
-                        for (Object o : al) {
-                            CfClasscontentkeyword cckw = new CfClasscontentkeyword();
-                            CfClasscontentkeywordPK cfClasscontentkeywordPK = new CfClasscontentkeywordPK();
-                            cfClasscontentkeywordPK.setClasscontentref(cfclasscontent.getId());
-                            cfClasscontentkeywordPK.setKeywordref(Long.parseLong(o.toString()));
-                            cckw.setCfClasscontentkeywordPK(cfClasscontentkeywordPK);
-                            cfclasscontentkeywordService.create(cckw);
+                        if (null != entity.getProperty("keywordset")) {
+                            // Delete corresponding keywordcontent entries
+                            List<CfClasscontentkeyword> keywordcontentdummy = cfclasscontentkeywordService.findByClassContentRef(cfclasscontent.getId());
+                            for (CfClasscontentkeyword keywordcontent : keywordcontentdummy) {
+                                cfclasscontentkeywordService.delete(keywordcontent);
+                            }
+                            ArrayList al = (ArrayList) entity.getProperty("keywordset").getValue();
+                            for (Object o : al) {
+                                CfClasscontentkeyword cckw = new CfClasscontentkeyword();
+                                CfClasscontentkeywordPK cfClasscontentkeywordPK = new CfClasscontentkeywordPK();
+                                cfClasscontentkeywordPK.setClasscontentref(cfclasscontent.getId());
+                                cfClasscontentkeywordPK.setKeywordref(Long.parseLong(o.toString()));
+                                cckw.setCfClasscontentkeywordPK(cfClasscontentkeywordPK);
+                                cfclasscontentkeywordService.create(cckw);
+                            }
                         }
                         
                         hibernateUtil.updateContent(cfclasscontent);
@@ -997,6 +1001,8 @@ public class EntityUtil {
                         CfKeywordlist keywordlist_search = cfkeywordlistService.findByName(name);
                         if (null == keywordlist_search) {
                             keywordlist.setName(name);
+                        } else {
+                            return false;
                         }
                             
                         // delete keywordlist entries
